+++ src/LiveComponent/tests/Integration/Twig/LiveComponentExtensionTest.php
@@ -26,5 +26,6 @@ final class LiveComponentExtensionTest extends KernelTestCase
 
         $this->assertStringContainsString('/_components/component1?props=%7B%22prop1%22:null,%22prop2%22:%222022-10-06T00:00:00%2B00:00%22,%22prop3%22:null,', $rendered);
         $this->assertStringContainsString('/alt/alternate_route?', $rendered);
+        $this->assertStringContainsString('http://localhost/_components/with_absolute_url?', $rendered);
     }
 }